[SPOT-286][REFACTOR] Report 서비스 컨트롤러 분리
